+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - Add get_shader_location
 - Add Shader#get_uniform_location
 - Add set_shader_values
+- Add Shader#set_value
+- Add Shader#set_values
 
 ## v0.3.13